+++ src/DataContainer/TinySliderSpreadContainer.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2019 Heimrich & Hannot GmbH
+ *
+ * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
+ * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ */
+
+
+namespace HeimrichHannot\TinySliderBundle\DataContainer;
+
+
+class TinySliderSpreadContainer
+{
+    const PALETTE_DEFAULT = 'default';
+    const PALETTE_PRESETCONFIG = 'presetConfig';
+    const PALETTE_GALLERY = 'gallery';
+    const PALETTE_CONTENT = 'tiny-slider';
+    const PALETTE_CONTENT_SLIDER_START = 'tiny-slider-content-start';
+    const PALETTE_CONTENT_SLIDER_END = 'tiny-slider-content-end';
+    const PALETTE_CONFIG_RESPONSIVE = 'responsive';
+    const PALETTE_CONFIG_BASE = 'flat';
+}